+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of the cookbook-openstack
 * Upgrading berkshelf from 2.0.18 to 3.1.5
 * Bump Chef gem to 11.16
 * Run trove-manage as user, not as root
+* Skip trove-manage run on SUSE, it is run by the init scripts already
 
 ## 9.0.1
 * Allow setting passwords via attributes by using the get_password method